Move Netty to separate module (#3551)
