+++ zio-http-netty/src/main/scala/zio/http/netty/client/ClientFailureHandler.scala
